Modernize SkFontStyle
Slowly rebuilding a better panose, the new scheme is that weight is
continuous from 0 (invisible) to 1000 (full coverage), width is a
continuous percentage with normal being 100, slant is a continuous angle
from -90 to 90 degrees, and italic is a continuous property from 0 (not
italic) to 1 (fully italic). This allows for the proper description of a
left leaning mosly italic font which is 98.6% as wide as normal and has
a weight of exactly 451. It is now possible to actually produce such
requests and fonts with some new variable fonts.

Bug: skia:6844

+++ src/svg/SkSVGDevice.cpp
@@ -643,7 +643,7 @@ void SkSVGDevice::AutoElement::addTextAttributes(const SkFont& font) {
     } else if (style.slant() == SkFontStyle::kOblique_Slant) {
         this->addAttribute("font-style", "oblique");
     }
-    int weightIndex = (SkTPin(style.weight(), 100, 900) - 50) / 100;
+    int weightIndex = (SkTPin((int)style.weight(), 100, 900) - 50) / 100;
     if (weightIndex != 3) {
         static constexpr const char* weights[] = {
             "100", "200", "300", "normal", "400", "500", "600", "bold", "800", "900"
